Fix initial window size and remember window pos
On startup, set the window's size to a more appropriate value so as not
to see the ugly scrollbars. Also, remember the window's location on
exit and reapply it on startup.

Author's note:

This should not be merged upstream! This is just a hasty custom patch
that I made that works well for my personal setup, where I use it with
my custom profile switcher.

I do not know whether the window size values used here only work well
on my machine, or whether they would also work well for everyone. On my
machine it is pixel perfect, such that the scroll bars don't appear but
it is also not any larger than it needs to be.

Also, for storing the window's location, it may be better to use
GUISettings now (i.e the .config file), rather than what is utilized
here (a .NET user.config file located in %LocalAppData%\grapher).

As I say, I just put this together for my own personal benefit.